+++ libs/nrlib/well/well.cpp
@@ -28,16 +28,15 @@
 #include "rmswell.hpp"
 
 #include "nrlib/iotools/logkit.hpp"
-//#include "nrlib/stormgrid/stormcontgrid.hpp"
-
 #include "nrlib/surface/surface.hpp"
+//#include "src/definitions.h"
 
 using namespace NRLib;
 
 Well::Well()
 {
   well_rmissing_ = -999.0;
-  well_imissing_  = -999;
+  well_imissing_ = -999;
 }
 
 
@@ -83,19 +82,20 @@ Well::ReadWell(const std::string              & file_name,
   if(file_name.find(".nwh",0) != std::string::npos) {
     NRLib::NorsarWell well(file_name);
     well_name_ = well.GetWellName();
-    cont_log_ = well.GetContLog();
-    n_data_ = well.GetNData();
-    disc_log_ = well.GetDiscLog();
-    read_ok = true;
+    cont_log_  = well.GetContLog();
+    n_data_    = well.GetNData();
+    disc_log_  = well.GetDiscLog();
+    read_ok    = true;
+
     // assume that facies logs from norsar wells are not used
     has_facies_log_ = false;
   }
   else if(file_name.find(".rms",0) != std::string::npos) {
     NRLib::RMSWell well(file_name);
     well_name_ = well.GetWellName();
-    n_data_ = well.GetNData();
-    cont_log_ = well.GetContLog();
-    disc_log_ = well.GetDiscLog();
+    n_data_    = well.GetNData();
+    cont_log_  = well.GetContLog();
+    disc_log_  = well.GetDiscLog();
 
     std::map<int, std::string> facies_map;
     if(well.HasDiscLog(facies_log))
@@ -297,7 +297,7 @@ int Well::CheckStormgrid(StormContGrid & stormgrid) const
     std::vector<double> y_pos = GetContLog().find("Y_pos")->second;
     std::vector<double> z_pos = GetContLog().find("TVD")->second;
 
-    for(int i=0; i < n_data_nonmissing_; i++) { //nd_
+    for(size_t i=0; i < n_data_nonmissing_; i++) { //nd_
 
       if(stormgrid.IsInside(x_pos[i], y_pos[i])) {
         insideArea = true;
@@ -329,3 +329,4 @@ int Well::CheckStormgrid(StormContGrid & stormgrid) const
   }
   return(error);
 }
+
